Split up the pages.rb file
