+++ src/lib/y2firewall/widgets/pages.rb
@@ -19,219 +19,6 @@
 # current contact information at www.suse.com.
 # ------------------------------------------------------------------------------
 
-require "yast"
-require "cwm/tabs"
-require "ui/service_status"
-require "y2firewall/firewalld"
-
-module Y2Firewall
-  module Widgets
-    module Pages
-      # A page for firewall service startup
-      class Startup < CWM::Page
-        # Constructor
-        #
-        # @param pager [CWM::TreePager]
-        def initialize(_pager)
-          textdomain "firewall"
-          Yast.import "SystemdService"
-
-          @service = Yast::SystemdService.find("firewalld")
-          # This is a generic widget in SLE15; may not be appropriate.
-          # For SLE15-SP1, use CWM::ServiceWidget
-          @status_widget = ::UI::ServiceStatus.new(@service)
-        end
-
-        # @macro seeAbstractWidget
-        def label
-          "Startup" # FIXME
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          VBox(
-            @status_widget.widget,
-            VStretch()
-          )
-        end
-      end
-
-      # A page for network interfaces, has {Interface} as subpages
-      class Interfaces < CWM::Page
-        # Constructor
-        #
-        # @param pager [CWM::TreePager]
-        def initialize(_pager)
-          textdomain "firewall"
-          @fw = Y2Firewall::Firewalld.instance
-          @fw.read # FIXME: when?
-        end
-
-        # @macro seeAbstractWidget
-        def label
-          "Interfaces" # FIXME
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          Label("TODO: List of interfaces here")
-        end
-      end
-
-      # A page for one network interface
-      class Interface < CWM::Page
-        # Constructor
-        #
-        # @param interface [Hash<String,String>] "id", "name" and "zone"
-        # @param pager [CWM::TreePager]
-        def initialize(interface, _pager)
-          textdomain "firewall"
-          @interface = interface
-          @sb = ZoneBox.new(interface)
-          self.widget_id = "ifc:" + label
-        end
-
-        # @macro seeAbstractWidget
-        def label
-          @interface["id"]
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          VBox(@sb)
-        end
-
-        # Selecting a zone to which an interface belongs
-        class ZoneBox < CWM::SelectionBox
-          # @param interface [Hash<String,String>] "id", "name" and "zone"
-          def initialize(interface)
-            @interface = interface
-            @zones = Y2Firewall::Firewalld.instance.zones
-          end
-
-          def label
-            format(_("Zone for Interface %s"), @interface["id"])
-          end
-
-          def items
-            @zones.map { |z| [z.name, z.name] }
-          end
-
-          def init
-            self.value = @interface["zone"]
-          end
-        end
-      end
-
-      # A page for firewall zones, has {Zone} as subpages
-      class Zones < CWM::Page
-        # Constructor
-        #
-        # @param pager [CWM::TreePager]
-        def initialize(_pager)
-          textdomain "firewall"
-          @fw = Y2Firewall::Firewalld.instance
-          @fw.read # FIXME: when?
-        end
-
-        # @macro seeAbstractWidget
-        def label
-          "Zones" # FIXME
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          Label("TODO: List of zones here")
-        end
-      end
-
-      # A page for a firewall zone
-      class Zone < CWM::Page
-        # Constructor
-        #
-        # @param zone [Y2Firewall::Firewalld::Zone]
-        # @param pager [CWM::TreePager]
-        def initialize(zone, pager)
-          textdomain "firewall"
-          @zone = zone
-          @pager = pager
-          self.widget_id = "z:" + zone.name
-        end
-
-        # @macro seeAbstractWidget
-        def label
-          @zone.name
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          VBox(
-            CWM::Tabs.new(
-              ServicesTab.new(@zone, @pager),
-              PortsTab.new
-            )
-          )
-        end
-      end
-
-      # A Tab for ports in a firewall zone
-      class PortsTab < CWM::Tab
-        def label
-          _("Ports")
-        end
-
-        def contents
-          VBox(
-            VStretch(),
-            HStretch()
-          )
-        end
-      end
-
-      # A Tab for services in a firewall zone
-      class ServicesTab < CWM::Tab
-        # Constructor
-        #
-        # @param zone [Y2Firewall::Firewalld::Zone]
-        # @param pager [CWM::TreePager]
-        def initialize(zone, _pager)
-          textdomain "firewall"
-          @zone = zone
-          @sb = ServiceBox.new(zone)
-          self.widget_id = "zs:" + zone.name
-        end
-
-        def label
-          _("Services")
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          VBox(@sb)
-        end
-
-        # A list of services in a firewall zone
-        class ServiceBox < CWM::MultiSelectionBox
-          # @param zone [Y2Firewall::Firewalld::Zone]
-          def initialize(zone)
-            @zone = zone
-          end
-
-          def label
-            # TRANSLATORS: %s is a zone name
-            format(_("Services for %s") % @zone.name)
-          end
-
-          def items
-            all_known_services = Y2Firewall::Firewalld.instance.api.services
-            all_known_services.map { |s| [s, s] }
-          end
-
-          def init
-            self.value = @zone.services
-          end
-        end
-      end
-    end
-  end
-end
+require "y2firewall/widgets/pages/interfaces"
+require "y2firewall/widgets/pages/startup"
+require "y2firewall/widgets/pages/zones"
